Refactor + fix content-length issue
This way need to generate a bit less code.

Also include the  Content-Length header always in POST requests made by the
client. This fixes an issue when we are serializing a message with no fields and
we have empty body. HttpClient does not include the Content-Length header for
us, so we'll just include it every time to make our requests work in all cases.